refactor: 코드 포맷팅
