+++ app/src/main/java/nextstep/jwp/http/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp;
+package nextstep.jwp.http;
 
 public enum ContentType {
     HTML("text/html;charset=utf-8"),
