+++ app/src/test/resources/static/responseFound.txt
@@ -1,0 +1,2 @@
+HTTP/1.1 302 Found
+Location: /index.html
